cleaned the detail individu page
