+++ src/pages/individual/[id]/Tabs/AutreContact.tsx
@@ -1,0 +1,6 @@
+import TableContact from '../Tables/TableContact'
+
+const AutreContact = () => {
+  return <TableContact />
+}
+export default AutreContact
