+++ src/pages/individual/[id]/Tabs/ListeRib.tsx
@@ -1,0 +1,7 @@
+import TableRib from '../Tables/TableRib'
+
+const ListeRib = () => {
+  return <TableRib />
+}
+
+export default ListeRib
