+++ src/pages/individual/[id]/Tabs/LoginWeb.tsx
@@ -1,0 +1,7 @@
+import TableLogin from '../Tables/TableLogin'
+
+const LoginWeb = () => {
+  return <TableLogin />
+}
+
+export default LoginWeb
